add IdeaFooter
